Keep Updated: receive tree clearing alerts text
The first button, "Draw an area" should link to the map with FORMA and the analysis tool turned on. If pressing back won't work once you're on the map, it should open in a new tab (so users can look at the instructions). 

The second button, "Select a concession or protected area" should link to the map with FORMA turned on. If pressing back won't work once you're on the map, it should open in a new tab (so users can look at the instructions). 

The last button, "Select a country or subnational region" should link to the country profiles. If pressing back won't work once you're on the map, it should open in a new tab (so users can look at the instructions). 

Let me know if I included too much text for the buttons. I can reduce it, and I might change this text anyway.