Adds a download link mailer on mobile
This commit introduces a download link mailer: when on a mobile device,
instead of linking directly to a Country download, we'll offer them the
ability to email themselves a link to the download.

* A DownloadView on the client decides if they are on mobile, and shows
a modal form if they are.
* GET /country/ISO/download redirects to the CartoDB download (user is
sent straight here if they're not on mobile).
* POST /country/ISO/download, given an email, sends the CartoDB link to
the given email.

The DownloadView on the client has been designed to be generic (it takes
any URL for a "download"), and so it can be used for mail notifications
for any type of download. The same goes for the mailer itself, however
the Countries controller is (unsurprisingly) only designed to create
Country downloads.