slight change to make new validation text visible
In my last change, I modified the cautions text for both gain and loss, but only gain showed up on the website. I think it is because there were extra lines/ spaces or something for loss but not gain. I deleted them and I believe this will fix the issue, but if not I'll follow up and have someone else take a look