Adds a Torque layer class
This class allows torque animation layers to be presented on the map.

It functions very similarly to the CartoDBLayerClass, and also inherits
from the OverlayLayerClass. This means that it can be used by adding a
new layer spec, as normal, and creating a layer that inherits from
TorqueLayerClass with the `table` and `column` options set.