Updating language to reflect AOI launch
I made a PR to change this from "Save this area to create a dashboard with a more in-depth analysis and receive email alerts about forest cover change" to "Save this area to receive email alerts about forest cover change. We are currently developing the ability to create dashboards for custom areas" text on March 24. Now that AOIs are launched, it's time to change it back.